refactor: move the defaultFormTemplates and registrar script into main resources folder
