+++ src/NextGen/DonationForm/resources/registrars/templates/index.ts
@@ -36,7 +36,7 @@ function extendTemplate<T extends keyof FormTemplates>(
         if (!window.givewp.form.templates[templateType].hasOwnProperty(template)) {
             throw Error(`${template} does not exist on ${templateType}`);
         }
-        
+
         window.givewp.form.templates[templateType][template] = templates[template];
     });
 }
